+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod cog;
 mod compression;
+mod cursor;
 pub mod error;
 mod ifd;
 mod partial_reads;
